Vegetable Harvest Project
This code is used to perform harvesting operations in two separate namespaces representing tomato and carrot gardens, as well as calculate the prices of the harvested products. The user is presented with options to choose either the tomato or carrot garden or calculate the prices of the products. After the user enters the prices, the respective harvesting operations are performed, and the total prices of the products are displayed on the screen. If there are no products to harvest or if the price is not entered, appropriate error messages are displayed. The user can choose the "Exit" option to terminate the program.