GameHub Project
This C++ program simulates a gaming platform on which different games can be played. The program allows the user to enter a username and then select a game from the main menu. There are three different game options: "Find the Synonym", "GuessMovie" and "Your Game".

In the "Find the Synonym" game, you try to find the synonym of a word. The program chooses a random word and asks you to guess its synonym. There is a certain number of tries and each correct guess earns you points.

The "GuessMovie" game is based on guessing the name of movies with different difficulty levels. The program selects a movie based on the selected difficulty level and asks you to guess the letters. You earn points when you guess the movie name completely with correct guesses.

The "Your Game" option is an option where the user can create his own game. There is no game logic for this option in the program and the user must implement his own game logic.

The program tracks the user's scores as they play games and eventually displays the user's name, total score, and games played.

This program shows how to create a basic gaming platform in C++ using class structure, inheritance, vectors and loops.